+++ GliglockTest/appCore/Test.cs
@@ -8,7 +8,7 @@
         }
         public Guid Id { get; set; }
         public string? Name { get; set; }
-        public int MaxMark { get; set; }
+        public ushort MaxMark { get; set; }
         public Guid TeacherId { get; set; }
         public List<Question> Questions { get; set; }
